:tada: Remove SWR
