+++ package.json
@@ -52,7 +52,6 @@
         "preact": "^10.4.1",
         "preact-render-to-string": "^5.1.8",
         "storeon": "^2.0.2",
-        "swr": "^0.2.0",
         "wouter": "^2.4.0"
     }
 }
