+++ src/layouts/dataProvider/types.ts
@@ -1,0 +1,8 @@
+import { FunctionComponent, VNode } from 'preact'
+
+interface DataProviderProps {
+    children: VNode
+}
+
+type DataProviderComponent = FunctionComponent<DataProviderProps>
+export default DataProviderComponent
