--- src/layouts/store/index.tsx
@@ -1,42 +1,0 @@
-import { h } from 'preact'
-import { useEffect } from 'preact/hooks'
-
-import useSwr from 'swr'
-import fetch from '../../libs/fetch'
-
-import { useStoreon } from 'storeon/preact'
-
-import StoreLayoutComponent from './types'
-
-const StoreLayout: StoreLayoutComponent = ({ children }) => {
-    let { data: dreamin } = useSwr(
-        [
-            'https://apollo-search-maidreamin.now.sh',
-            'getMenu',
-            `
-                query getMenu {
-                    getMenu {
-                        name {
-                            th
-                            en
-                            jp
-                        }
-                        subMenu
-                        price
-                    }
-                }
-            `
-        ],
-        fetch
-    )
-
-    const { dispatch } = useStoreon()
-
-    useEffect(() => {
-        dispatch('UPDATE_MENU', typeof dreamin !== 'undefined' ? dreamin : {})
-    }, [dreamin])
-
-    return children
-}
-
-export default StoreLayout
