--- src/layouts/store/types.ts
@@ -1,8 +1,0 @@
-import { FunctionComponent, VNode } from 'preact'
-
-interface StoreLayoutProps {
-    children: VNode
-}
-
-type StoreLayoutComponent = FunctionComponent<StoreLayoutProps>
-export default StoreLayoutComponent
