+++ src/libs/api.ts
@@ -1,22 +1,24 @@
 import { Dreamin } from '../store/dreamin/types'
 
-export default async (...args: string[]): Promise<Dreamin[]> => {
+const getMenu = async (operationName: string, host: string, query: string): Promise<Dreamin[]> => {
     if (typeof window === 'undefined') return []
 
     const fetch = await require('isomorphic-unfetch')
 
-    const response = await fetch(args[0], {
+    const response = await fetch(host, {
         method: 'POST',
         headers: { 'Content-Type': 'application/json' },
         mode: 'cors',
         body: JSON.stringify({
-            operationName: args[1],
-            query: args[2],
+            operationName,
+            query,
             variables: []
         })
     })
 
     const data = await response.json()
 
-    return data.data[args[1]] as Dreamin[]
+    return data.data[operationName] as Dreamin[]
 }
+
+export default getMenu
