+++ src/styles/404.styl
@@ -11,6 +11,7 @@
     & > .illust
         display block
         width 300px
+        height 208.25px
         margin 0 0 40px
 
     & > .title
