+++ src/styles/menu-list.styl
@@ -17,7 +17,12 @@
     & > .illust
         display block
         width 80%
+        height 251px
         margin 20px auto 40px auto
+
+        @media screen and (max-width 568px)
+            width 280px
+            height 200px
 
     & > .title
         width 100%
