+++ src/template.html
@@ -101,6 +101,9 @@
             media="(device-width: 1024px) and (device-height: 1366px) and (-webkit-device-pixel-ratio: 2)"
             rel="apple-touch-startup-image"
         />
+
+        <link rel="preload" as="image" href="/assets/illustrations/loading.svg">
+        <link rel="preconnect" href="https://apollo-search-maidreamin.now.sh">
         <% preact.headEnd %>
     </head>
     <body>
